Abstract code to own files
