+++ src/App.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 
 import { BrowserRouter as Router } from 'react-router-dom';
 
-import { LightboxProvider } from './hooks/LightboxProvider';
 import './index.css';
+import { LightboxProvider } from './providers/LightboxProvider';
 import AppRoutes from './routes';
 
 const App = () => {
