+++ src/components/galleryPage/GalleryItem.jsx
@@ -6,11 +6,12 @@
  *
  * @param {string} src - The source URL of the image
  * @param {string} alt - The alternative text for the image
+ * @param {string} caption - The caption text to display below the image
  * @param {string} aspectClass - The CSS class for the image aspect ratio
  * @param {function} onClick - The function to call when the image is clicked
  * @returns {React.ReactNode} - The GalleryItem component
  */
-const GalleryItem = ({ src, alt, aspectClass, onClick }) => (
+const GalleryItem = ({ src, alt, caption, aspectClass, onClick }) => (
   <div className="group w-full h-full flex flex-col">
     <div
       className={`relative w-full h-full overflow-hidden shadow-sm ${aspectClass}`}
@@ -23,9 +24,15 @@ const GalleryItem = ({ src, alt, aspectClass, onClick }) => (
         onClick={onClick}
       />
     </div>
-    <p className="mt-2 mb-4 text-lg text-center text-gray-600">
-      Lorem ipsum dolor sit amet
-    </p>
+    {caption && (
+      <div className="w-full mt-2 mb-4">
+        <div className="bg-white py-2 px-3 rounded">
+          <p className="text-sm md:text-base text-center font-semibold text-gray-900">
+            {caption}
+          </p>
+        </div>
+      </div>
+    )}
   </div>
 );
